requirements.txt
mcp>=1.0.0
requests>=2.31.0
requests-oauthlib>=1.3.1
python-dotenv>=1.0.0
